Make /diff the front door
The hero now leads with "Read a diff" and demotes account creation to a
secondary action; the curated incident diffs move to
src/features/incident-diffs/ and render on the landing page under the
hand-authored scan preview; shared public reports close with the same
conversion block shape as /diff.
