+++ docs/design.md
@@ -409,3 +409,4 @@ The canonical "labeled control with metadata" row pattern. Re-use this compositi
 | 2026-07-22 | Diff scroll/hidden-line pass: overview strip maps scroll space and gains a viewport thumb; gap expanders offer capped `show all`; scroll pane becomes a keyboard region. | The overview strip positioned markers over logical file lines while the scrollbar operates over the collapsed display, so a 30k-line file showed markers at opposite ends for changes one screen apart — it now maps the rendered rows (a collapsed gap counts as its one expander row) and a `bg-ink/15` thumb tracks the viewport, keeping ink structural per "color = signal". Gap expanders add a one-click `show all` capped at 5,000 hidden rows so bulk reveal can't reintroduce the megabyte-scale freeze. The 560px pane was mouse-only: it's now a labeled `role="region"` with `tabIndex=0`, an inset accent-soft focus ring (outer halos clip against the shell), and contained overscroll. Scroll anchoring is off so expansion never jumps in any browser, and late-arriving findings no longer reset the scroll position (the reset key is file identity + content only).                                                                                                                                                                                             |
 | 2026-07-22 | Component-gap pass: keyboard/AT fixes across overlays and the tree, primitives pinned, spec drift reconciled.                                                            | A cross-component audit found the gaps clustered in keyboard/AT support and primitive drift. Fixed: CloseButton no longer suppresses the global focus ring; Menu returns focus to its trigger on item activation and closes on Tab-out; critical Toasts persist until dismissed and timed ones pause on hover/focus; Select forwards rest props so selects can be named (VersionPicker now uses `label/for`); Dialogs lock background scroll, ignore drag-to-backdrop clicks, and focus their first real control; FileTree rows carry `aria-pressed`; PageShell gains a banner landmark + skip link; SectionLabel renders a real `h2` with an `aside` slot and `MonoLabel` pins the bare label; SeverityBar is a div stack (last non-sanctioned SVG removed) with its name label at the 11px floor; Alert discs use saturated tokens; small primary/danger Buttons hold 13px. Spec updated to match reality where the code had deliberately outgrown it: highlight grammar list, FileTree status conveyance (no trailing status Badge), and the 20px indent now measured spacer+gap. |
 | 2026-07-22 | `Package diff` linked from both the marketing header and the site footer.                                                                                                | The public diff tool is the no-account entry point to the product but had no navigation into it — visitors reached `/diff` only via a direct link. It now sits in `MarketingHeaderActions` (landing/docs/diff headers) as the discovery path, and in the site footer so app routes, whose headers carry dashboard actions instead of the marketing nav, still reach it. Both use the same label so the two surfaces read as one destination; `Docs` already appeared in both, so the duplication follows existing precedent rather than introducing a second nav vocabulary. No active-state treatment was added — the marketing header has never marked the current page, and adding it for one link only would be inconsistent.                                                                                                                                                                                                                                                                                                                                                    |
+| 2026-08-08 | `/diff` becomes the front door: hero leads with `Read a diff`, live incident diffs join the landing, public reports gain a conversion close.                             | The zero-setup surface (`/diff`) was the product's strongest proof but reachable only from a ghost header link, while the hero's sole CTA asked for signup before showing any value. The hero now leads with `Read a diff` (primary) and demotes `Create account`/`Sign in` to secondary/ghost. The curated incident diffs move into `src/features/incident-diffs/` (shared by the `/diff` landing and the marketing landing per the two-page rule) and render under the hand-authored scan preview with copy naming the preview as hand-authored and the incidents as live. Shared public reports — warm traffic from existing users — close with the same conversion block shape as `/diff` (`Before it shipped`), linking `Create account` and `Diff a package`; the block sits after `Verify this report` so proof outranks pitch.                                                                                                                                                                                                                                               |
