+++ src/pages/PublicReport/index.tsx
@@ -10,7 +10,7 @@ import { FindingCard } from "../../components/FindingCard";
 import { LoadingState } from "../../components/Loading";
 import { PageShell } from "../../components/PageShell";
 import { LinkButton } from "../../components/Button";
-import { EmptyLine, MonoDetail, SectionLabel } from "../../components/Typography";
+import { EmptyLine, MonoDetail, Muted, SectionLabel } from "../../components/Typography";
 import { verdictTextClass } from "../../features/review/verdict";
 import { MarketingHeaderActions } from "../MarketingHeaderActions";
 import { useAuthedSession } from "../useAuthedSession";
@@ -289,6 +289,25 @@ export default function PublicReportPage() {
             download
           >
             Download report JSON
+          </LinkButton>
+        </div>
+      </section>
+
+      <section class="flex flex-col gap-3 pt-3">
+        <SectionLabel as="p">Before it shipped</SectionLabel>
+        <h2 class="text-2xl font-semibold tracking-[-0.015em] m-0">
+          This release was reviewed while it could still be stopped.
+        </h2>
+        <Muted class="m-0 text-[14px] leading-[1.65] max-w-[680px]">
+          The publisher held this release — an npm staged publish or a GitHub-gated job — and read
+          this report before letting it ship. Drydock runs the same review on every version you
+          publish; the maintainer keeps the final decision. You can also diff any published npm or
+          PyPI package without an account.
+        </Muted>
+        <div class="flex flex-wrap gap-3 mt-1">
+          <LinkButton href="/register">Create account</LinkButton>
+          <LinkButton href="/diff" variant="secondary">
+            Diff a package
           </LinkButton>
         </div>
       </section>
